ReportControllerIntegrationTests better organized (split in 3 classes)   and removed outdated comment
