+++ src/test/java/edu/rit/se/testsmells/testsmell/ReportControllerIntegrationTest.java
@@ -44,45 +44,6 @@ class ReportControllerIntegrationTest {
     }
 
     @Test
-    void testMultipleOutput_METHOD() throws IOException {
-        List<ReportController.ReportGranularity> outputs = Arrays.asList(ReportController.ReportGranularity.METHOD, ReportController.ReportGranularity.CLASS, ReportController.ReportGranularity.FILE);
-        sut = new ReportController(csvWriter, outputs);
-        sut.report(Collections.singletonList(file));
-
-        outputFile = new File("METHOD_" + csvWriter.getSuffix());
-        assertTrue(outputFile.exists(), "Method output file missing!");
-        List<String> fileContent = new BufferedReader(new FileReader(outputFile)).lines().collect(Collectors.toList());
-        assertEquals(3, fileContent.size(), "File with unexpected size received! File's content is: " + System.lineSeparator() + String.join(System.lineSeparator(), fileContent) + System.lineSeparator());
-    }
-
-    @Test
-    void testMultipleOutput_CLASS() throws IOException {
-        List<ReportController.ReportGranularity> outputs = Arrays.asList(ReportController.ReportGranularity.METHOD, ReportController.ReportGranularity.CLASS, ReportController.ReportGranularity.FILE);
-        sut = new ReportController(csvWriter, outputs);
-        sut.report(Collections.singletonList(new TestFile("ConstructorInitializationProject", "/ConstructorInitialization/src/test/java/org/briarproject/bramble/crypto/TagEncodingTest.java", "")));
-
-        outputFile = new File("CLASS_" + csvWriter.getSuffix());
-        assertTrue(outputFile.exists(), "Class output file missing!");
-        List<String> fileContent = new BufferedReader(new FileReader(outputFile)).lines().collect(Collectors.toList());
-        assertEquals(0, fileContent.size(), "File with unexpected size received! File's content is: " + System.lineSeparator() + String.join(System.lineSeparator(), fileContent) + System.lineSeparator());
-
-    }
-
-    @Test
-    void testMultipleOutput_FILE() throws IOException {
-        List<ReportController.ReportGranularity> outputs = Arrays.asList(ReportController.ReportGranularity.METHOD, ReportController.ReportGranularity.CLASS, ReportController.ReportGranularity.FILE);
-        sut = new ReportController(csvWriter, outputs);
-        sut.report(Collections.singletonList(file));
-
-        outputFile = new File("FILE_" + csvWriter.getSuffix());
-        assertTrue(outputFile.exists(), "File output file missing!");
-        List<String> fileContent = new BufferedReader(new FileReader(outputFile)).lines().collect(Collectors.toList());
-        assertEquals(2, fileContent.size(), "File with unexpected size received! File's content is: " + System.lineSeparator() + String.join(System.lineSeparator(), fileContent) + System.lineSeparator());
-
-
-    }
-
-    @Test
     void testReport() {
         sut = new ReportController(csvWriter, Collections.singletonList(ReportController.ReportGranularity.FILE));
         assertDoesNotThrow(() -> sut.report(Collections.singletonList(file)));
@@ -207,42 +168,3 @@ class ReportControllerIntegrationTest {
     }
 }
 
-class ReportControllerClassIntegrationTests {
-    private File outputFile;
-    private TestFile file;
-    private ReportController sut;
-    private CSVWriter csvWriter;
-    private List<AbstractSmell> smells = Arrays.asList(new AssertionRoulette(), new ConditionalTestLogic(), new ConstructorInitialization(), new DefaultTest(), new EmptyTest(), new ExceptionCatchingThrowing(), new GeneralFixture(), new MysteryGuest(), new PrintStatement(), new RedundantAssertion(), new SensitiveEquality(), new VerboseTest(), new SleepyTest(), new EagerTest(), new LazyTest(), new DuplicateAssert(), new UnknownTest(), new IgnoredTest(), new ResourceOptimism(), new MagicNumberTest(), new DependentTest());
-    private String appName = "ConstructorInitializationProject";
-    private String testFilePath = "/ConstructorInitialization/src/test/java/org/briarproject/bramble/crypto/TagEncodingTest.java";
-    private String productionFilePath = "";
-
-    @BeforeEach
-    void setUp() throws IOException {
-        TestSmellDetector testSmellDetector = TestSmellDetector.createTestSmellDetector();
-        smells.forEach(testSmellDetector::addDetectableSmell);
-        file = new TestFile(appName, testFilePath, productionFilePath);
-        csvWriter = CSVWriter.createResultsWriter();
-        testSmellDetector.detectSmells(file);
-    }
-
-    @AfterEach
-    void tearDown() {
-        //noinspection ResultOfMethodCallIgnored
-        outputFile.delete();
-    }
-
-
-    @Test
-    void testCLASSOutput() throws IOException {
-        List<ReportController.ReportGranularity> outputs = Arrays.asList(ReportController.ReportGranularity.CLASS);
-        sut = new ReportController(csvWriter, outputs);
-        sut.report(Collections.singletonList(file));
-
-        outputFile = new File("CLASS_" + csvWriter.getSuffix());
-        assertTrue(outputFile.exists(), "Class output file missing!");
-        List<String> fileContent = new BufferedReader(new FileReader(outputFile)).lines().collect(Collectors.toList());
-        assertEquals(2, fileContent.size(), "File with unexpected size received! File's content is: " + System.lineSeparator() + String.join(System.lineSeparator(), fileContent) + System.lineSeparator());
-
-    }
-}
